services/worker-grants/requirements.txt
livekit-agents>=1.2,<1.3
livekit-plugins-openai>=1.2,<1.3
livekit-plugins-elevenlabs>=1.2,<1.3
livekit-plugins-simli>=1.2,<1.3
python-dotenv>=1.0,<2.0
httpx>=0.24.0,<1.0
